refactor: Pull ORS HTTP Client out to its own module
