+++ mix.exs
@@ -69,6 +69,7 @@ defmodule Skate.MixProject do
       {:jason, "~> 1.0"},
       {:lcov_ex, "~> 0.2", only: [:dev, :test], runtime: false},
       {:logster, "~> 1.0"},
+      {:mox, "~> 1.1.0"},
       {:oban, "~> 2.15"},
       {:phoenix_ecto, "~> 4.0"},
       {:phoenix_html, "~> 3.0"},
